requirements.txt
pandas
requests
python-dotenv
supabase
psycopg2-binary
beautifulsoup4
